requirements.txt
langchain==0.1.13
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.33
openai==1.25.2
faiss-cpu==1.7.4
altair==4
tiktoken==0.5.2
langchain-openai == 0.1.6
transformers == 4.40.1
pillow == 10.3.0
torch == 2.3.0
torchvision == 0.18.0
youtube-transcript-api == 0.6.2
chromadb == 0.3.29
opencv-python-headless == 4.9.0.80

